+++ id_us.h
@@ -112,7 +112,7 @@ extern	boolean		Button0,Button1,
 					CursorBad;
 extern	Sint16			CursorX,CursorY;
 
-extern	void		(*USL_MeasureString)(const char*,Uint16 *,Uint16 *),
+extern	void		(*USL_MeasureString)(const char*, int*, int*),
 					(*USL_DrawString)(const char*);
 
 extern	boolean		(*USL_SaveGame)(Sint16),(*USL_LoadGame)(Sint16);
